+++ lumen/tests/ai/conftest.py
@@ -11,6 +11,7 @@ except ModuleNotFoundError:
 from pydantic import BaseModel
 
 from lumen.ai.llm import Llm, Message
+from lumen.sources.duckdb import DuckDBSource
 
 
 def pytest_collection_modifyitems(config, items):
@@ -61,8 +62,6 @@ def llm():
 
 @pytest.fixture
 def tiny_source():
-    from lumen.sources.duckdb import DuckDBSource
-
     return DuckDBSource(tables={
     'tiny': """
         SELECT * FROM (
