+++ lumen/tests/conftest.py
@@ -1,5 +1,4 @@
 import asyncio
-import contextlib
 import os
 import tempfile
