+++ lumen/tests/sources/conftest.py
@@ -2,7 +2,6 @@ import tempfile
 
 from pathlib import Path
 
-import duckdb
 import pandas as pd
 import pytest
 
@@ -26,25 +25,3 @@ def source_tables():
     }
     yield tables
     pd.set_option('mode.string_storage', string)
-
-
-@pytest.fixture
-def duckdb_file_source():
-    """
-    Create a temporary DuckDB database file with a test table.
-
-    Yields the DuckDBSource and cleans up after the test.
-    """
-    with tempfile.TemporaryDirectory() as tmpdir:
-        db_path = Path(tmpdir) / 'test.duckdb'
-        conn = duckdb.connect(str(db_path))
-        conn.execute('CREATE TABLE test (id INTEGER, name VARCHAR)')
-        conn.execute("INSERT INTO test VALUES (1, 'Alice'), (2, 'Bob'), (3, 'Charlie')")
-        conn.close()
-
-        source = DuckDBSource(
-            uri=str(db_path),
-            tables=['test'],
-        )
-        yield source
-        source.close()
